+++ src/screens/WalletScreen/AppTour/Boxes/Steps/Step4.tsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import { View } from 'react-native';
+import { Text, Icon } from '@components';
+
+export const Step4 = () => (
+	<>
+		<View style={{ height: 25, marginBottom: 8, flexDirection: 'row' }}>
+			<Icon
+				name="sendStroke"
+				size={20}
+				color="cta1"
+				style={{ marginRight: 8 }}
+			/>
+			<Text
+				type="p" // Change to tMedium after new values come from merge.
+				weight="bold"
+			>
+				Send
+			</Text>
+		</View>
+		<Text width={237} type="a" marginBottom={16}>
+			Send tokens to a another wallet or to an exchange like Binance or Coinbase.
+		</Text>
+
+		<View style={{ height: 25, flexDirection: 'row' }}>
+			<Icon
+				name="exchangeStroke"
+				size={20}
+				color="cta1"
+				style={{ marginRight: 8 }}
+			/>
+			<Text
+				type="p" // Change to tMedium after new values come from merge.
+				weight="bold"
+			>
+				Exchange
+			</Text>
+		</View>
+		<Text width={237} type="a">
+			Swap between tokens.
+		</Text>
+	</>
+);
