+++ src/screens/WalletScreen/AppTour/Boxes/Steps/index.ts
@@ -1,0 +1,15 @@
+import { Step0 } from './Step0';
+import { Step1 } from './Step1';
+import { Step2 } from './Step2';
+import { Step3 } from './Step3';
+import { Step4 } from './Step4';
+import { Step5 } from './Step5';
+
+export {
+	Step0,
+	Step1,
+	Step2,
+	Step3,
+	Step4,
+	Step5
+};
